+++ internal/oci/containerd.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"net/url"
 	"path"
+	"sort"
 	"strings"
 
 	"github.com/containerd/containerd"
@@ -21,27 +22,67 @@ import (
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/spf13/afero"
 	"github.com/xenitab/pkg/channels"
-
-	"github.com/xenitab/spegel/internal/header"
+	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1"
 )
 
 type Containerd struct {
-	client      *containerd.Client
-	listFilter  string
-	eventFilter string
+	client             *containerd.Client
+	platform           platforms.MatchComparer
+	listFilter         string
+	eventFilter        string
+	runtimeClient      runtimeapi.RuntimeServiceClient
+	registryConfigPath string
 }
 
-func NewContainerd(sock, namespace string, registries []url.URL) (*Containerd, error) {
+func NewContainerd(sock, namespace, registryConfigPath string, registries []url.URL) (*Containerd, error) {
 	client, err := containerd.New(sock, containerd.WithDefaultNamespace(namespace))
 	if err != nil {
 		return nil, fmt.Errorf("could not create containerd client: %w", err)
 	}
 	listFilter, eventFilter := createFilters(registries)
+	runtimeClient := runtimeapi.NewRuntimeServiceClient(client.Conn())
 	return &Containerd{
-		client:      client,
-		listFilter:  listFilter,
-		eventFilter: eventFilter,
-	}, err
+		client:             client,
+		platform:           platforms.Default(),
+		listFilter:         listFilter,
+		eventFilter:        eventFilter,
+		runtimeClient:      runtimeClient,
+		registryConfigPath: registryConfigPath,
+	}, nil
+}
+
+func (c *Containerd) Verify(ctx context.Context) error {
+	ok, err := c.client.IsServing(ctx)
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return fmt.Errorf("could not reach Containerd service")
+	}
+	resp, err := c.runtimeClient.Status(ctx, &runtimeapi.StatusRequest{Verbose: true})
+	if err != nil {
+		return err
+	}
+	str, ok := resp.Info["config"]
+	if !ok {
+		return fmt.Errorf("could not get config data from info response")
+	}
+	cfg := &struct {
+		Registry struct {
+			ConfigPath string `json:"configPath"`
+		} `json:"registry"`
+	}{}
+	err = json.Unmarshal([]byte(str), cfg)
+	if err != nil {
+		return err
+	}
+	if cfg.Registry.ConfigPath == "" {
+		return fmt.Errorf("Containerd registry config path needs to be set for mirror configuration to take effect")
+	}
+	if cfg.Registry.ConfigPath != c.registryConfigPath {
+		return fmt.Errorf("Containerd registry config path is %s but needs to be %s for mirror configuration to take effect", cfg.Registry.ConfigPath, c.registryConfigPath)
+	}
+	return nil
 }
 
 func (c *Containerd) Subscribe(ctx context.Context) (<-chan Image, <-chan error) {
@@ -89,8 +130,9 @@ func (c *Containerd) ListImages(ctx context.Context) ([]Image, error) {
 
 func (c *Containerd) GetImageDigests(ctx context.Context, img Image) ([]string, error) {
 	keys := []string{}
-	platform := platforms.Default()
 	err := images.Walk(ctx, images.HandlerFunc(func(ctx context.Context, desc ocispec.Descriptor) ([]ocispec.Descriptor, error) {
+		keys = append(keys, desc.Digest.String())
+
 		b, err := content.ReadBlob(ctx, c.client.ContentStore(), desc)
 		if err != nil {
 			return nil, err
@@ -106,14 +148,32 @@ func (c *Containerd) GetImageDigests(ctx context.Context, img Image) ([]string, 
 			if err := json.Unmarshal(b, &idx); err != nil {
 				return nil, err
 			}
-			for _, manifest := range idx.Manifests {
-				if !platform.Match(*manifest.Platform) {
+			var descs []ocispec.Descriptor
+			for _, m := range idx.Manifests {
+				if !c.platform.Match(*m.Platform) {
 					continue
 				}
-				keys = append(keys, manifest.Digest.String())
-				return []ocispec.Descriptor{manifest}, nil
+				descs = append(descs, m)
 			}
-			return nil, fmt.Errorf("could not find platform architecture in manifest: %v", desc.Digest)
+
+			if len(descs) == 0 {
+				return nil, fmt.Errorf("could not find platform architecture in manifest: %v", desc.Digest)
+			}
+
+			// Platform matching is a bit weird in that multiple platforms can match.
+			// There is however a "best" match that should be used.
+			// This logic is used by Containerd to determine which layer to pull so we should use the same logic.
+			sort.SliceStable(descs, func(i, j int) bool {
+				if descs[i].Platform == nil {
+					return false
+				}
+				if descs[j].Platform == nil {
+					return true
+				}
+				return c.platform.Less(*descs[i].Platform, *descs[j].Platform)
+			})
+
+			return []ocispec.Descriptor{descs[0]}, nil
 		case images.MediaTypeDockerSchema2Manifest, ocispec.MediaTypeImageManifest:
 			var manifest ocispec.Manifest
 			if err := json.Unmarshal(b, &manifest); err != nil {
@@ -123,7 +183,6 @@ func (c *Containerd) GetImageDigests(ctx context.Context, img Image) ([]string, 
 			for _, layer := range manifest.Layers {
 				keys = append(keys, layer.Digest.String())
 			}
-			// TODO: In the images.Manifest implementation there is a platform check that I do not understand
 			return nil, nil
 		}
 		return nil, fmt.Errorf("unexpected media type %v for digest: %v", ud.MediaType, desc.Digest)
@@ -134,7 +193,6 @@ func (c *Containerd) GetImageDigests(ctx context.Context, img Image) ([]string, 
 	if len(keys) == 0 {
 		return nil, fmt.Errorf("no image digests found")
 	}
-	keys = append(keys, img.Digest.String())
 	return keys, nil
 }
 
@@ -251,20 +309,11 @@ func hostsFileContent(registryURL url.URL, mirrorURLs []url.URL) string {
 		server = "https://registry-1.docker.io"
 	}
 	content := fmt.Sprintf(`server = "%s"`, server)
-	for i, mirrorURL := range mirrorURLs {
-		content = fmt.Sprintf(`%[1]s
+	for _, mirrorURL := range mirrorURLs {
+		content = fmt.Sprintf(`%s
 
-[host."%[3]s"]
-  capabilities = ["pull", "resolve"]
-[host."%[3]s".header]
-  %[4]s = ["%[2]s"]
-  %[5]s = ["true"]`, content, registryURL.String(), mirrorURL.String(), header.RegistryHeader, header.MirrorHeader)
-
-		// We assume first mirror registry is local. All others are external.
-		if i != 0 {
-			content = fmt.Sprintf(`%s
-  %s = ["true"]`, content, header.ExternalHeader)
-		}
+[host."%s"]
+  capabilities = ["pull", "resolve"]`, content, mirrorURL.String())
 	}
 	return content
 }